implement better support for floats
this change uses `fround` when available to make a better decision on
if a number is a 32 or 64 bit floating point number.  This means that
we can more safely use floats in msgpack without losing precision.

In the case where `Math.fround` is not defined (I'M LOOKIN AT YOU, IE10),
this code defaults to using a float64 so that we don't accidentally lose
precision like we did with the previous implementation